+++ mealz-shared-analytics/src/commonMain/kotlin/mealz/ai/PlausibleEvent.kt
@@ -5,9 +5,9 @@ import kotlin.js.JsName
 
 @JsExport
 @JsName("PlausibleEvent")
-data class PlausibleEvent(
+class PlausibleEvent(
     val name: String,
     val url: String,
     val domain: String,
-    val props: PlausibleProps
+    val props: Any
 )